heimdal_data/requirements.txt
# API Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.1

# Data Collection
tweepy>=4.4.0
facebook-sdk==3.1.0
pytrends>=4.7.3
pandas>=1.3.2
requests>=2.26.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.2
apscheduler>=3.8.1
loguru>=0.5.3

# Optional - OpenAI for trend analysis
openai>=0.27.0
